// types.ts
export enum UserRole {
  ADMIN = 'Administrator',
  CUSTOMER = 'Customer', // Updated from BUYER
  MANUFACTURER = 'Manufacturer',
}